refactoring testing files and removing redundant imports
