+++ src/test/java/DataBaseTest/DataBaseParserTest.java
@@ -1,3 +1,5 @@
+package DataBaseTest;
+
 import DataBase.DataBaseParser;
 import Entities.*;
 import org.json.simple.JSONObject;
@@ -5,7 +7,6 @@ import org.json.simple.parser.JSONParser;
 import org.json.simple.parser.ParseException;
 import org.junit.jupiter.api.Assertions;
 import org.junit.jupiter.api.Test;
-
 import java.util.ArrayList;
 import java.util.Date;
