+++ src/test/java/DataBaseTest/DataBaseTest.java
@@ -1,3 +1,5 @@
+package DataBaseTest;
+
 import DataBase.DataBase;
 import org.junit.jupiter.api.Assertions;
 import org.junit.jupiter.api.Test;
@@ -5,7 +7,7 @@ import Entities.*;
 import java.util.ArrayList;
 import java.util.Date;
 public class DataBaseTest {
-    ArrayList<Double> priceData = new ArrayList();
+    ArrayList<Double> priceData = new ArrayList<>();
     ArrayList<Date> priceDate = new ArrayList<>();
     String testDate = "Fri Nov 18 01:04:05 EST 2022";
     Item plushie = new Item("Plushie", 40.99, 30.00, "www.amazon.com/plushie",
