+++ src/test/java/EntitiesTest/ItemTest.java
@@ -1,8 +1,9 @@
+package EntitiesTest;
+
 import Entities.Item;
 import org.junit.jupiter.api.Assertions;
 import org.junit.jupiter.api.Test;
 import java.io.IOException;
-import java.util.ArrayList;
 
 public class ItemTest {
 //    /**
@@ -48,7 +49,7 @@ public class ItemTest {
 
         double newPrice = priceUpdateTestItem.getProductPrice();
 
-        Assertions.assertEquals(true, initialPrice != newPrice);
+        Assertions.assertTrue(initialPrice != newPrice);
     }
 
     /**
@@ -64,7 +65,7 @@ public class ItemTest {
                         "Ships in Certified Frustration-free Packaging", new String[]{"computer accesssories", "Tech", "office"}, 0, 0, "imageurl");
         // Test 1:  Set and Get Name
         TestItem.setName("AmazonBasics Wired Office Keyboard");
-        Assertions.assertEquals(true, "AmazonBasics Wired Office Keyboard".equals(TestItem.getProductName()));
+        Assertions.assertEquals("AmazonBasics Wired Office Keyboard", TestItem.getProductName());
 
 
     }
@@ -81,7 +82,7 @@ public class ItemTest {
                         "Backed by One-year Amazon Basics Warranty\n" +
                         "Ships in Certified Frustration-free Packaging", new String[]{"computer accesssories", "Tech", "office"}, 0, 0, "imageurl");
 
-        Assertions.assertEquals(true, 20.00 == TestItem.getProductPrice());
+        Assertions.assertEquals(20.00, TestItem.getProductPrice());
     }
 
     /**
@@ -97,7 +98,7 @@ public class ItemTest {
                         "Ships in Certified Frustration-free Packaging", new String[]{"computer accesssories", "Tech", "office"}, 0, 0, "imageurl");
 
         TestItem.setDesiredPrice(17.00);
-        Assertions.assertEquals(true, 17.00 == TestItem.getProductDesiredPrice());
+        Assertions.assertEquals(17.00, TestItem.getProductDesiredPrice());
     }
 
     /**
@@ -113,7 +114,7 @@ public class ItemTest {
                         "Ships in Certified Frustration-free Packaging", new String[]{"computer accesssories", "Tech", "office"}, 0, 0, "imageurl");
         String testUrl = "https://www.amazon.ca/AmazonBasics-KU-0833-Wired-Keyboard/dp/B005EOWBHC/ref=sr_1_6?crid=LXQRVB06NTVV&keywords=keyboard&qid=1668040664&qu=eyJxc2MiOiI3LjM4IiwicXNhIjoiNi42NSIsInFzcCI6IjUuOTMifQ%3D%3D&sprefix=keyboard%2Caps%2C90&sr=8-6&th=1";
 
-        Assertions.assertEquals(true, testUrl.equals(TestItem.getProductURL()));
+        Assertions.assertEquals(testUrl, TestItem.getProductURL());
     }
 
     /**
@@ -129,7 +130,7 @@ public class ItemTest {
                         "Ships in Certified Frustration-free Packaging", new String[]{"computer accesssories", "Tech", "office"}, 0, 0, "imageurl");
         String testImageUrl = "imageurl";
 
-        Assertions.assertEquals(true, testImageUrl.equals(TestItem.getProductImageURL()));
+        Assertions.assertEquals(testImageUrl, TestItem.getProductImageURL());
     }
 
     /**
@@ -146,7 +147,7 @@ public class ItemTest {
         String newDescription = "This is a new set item description for AmazonBasics Wired Office Keyboard";
 
         TestItem.setProductDescription(newDescription);
-        Assertions.assertEquals(true, newDescription.equals(TestItem.getProductDescription()));
+        Assertions.assertEquals(newDescription, TestItem.getProductDescription());
     }
 
     /**
@@ -162,7 +163,7 @@ public class ItemTest {
                         "Ships in Certified Frustration-free Packaging", new String[]{"computer accesssories", "Tech", "office"}, 0, 0, "imageurl");
 
         TestItem.setReviewCount(30);
-        Assertions.assertEquals(true, 30 == TestItem.getReviewCount());
+        Assertions.assertEquals(30, TestItem.getReviewCount());
     }
 
     /**
@@ -178,6 +179,6 @@ public class ItemTest {
                         "Ships in Certified Frustration-free Packaging", new String[]{"computer accesssories", "Tech", "office"}, 0, 0, "imageurl");
 
         TestItem.setReviewStars(4.2);
-        Assertions.assertEquals(true, 4.2 == TestItem.getReviewStars());
+        Assertions.assertEquals(4.2, TestItem.getReviewStars());
     }
 }
