+++ src/test/java/ExternallnterfaceTest/ItemSearchTest.java
@@ -1,8 +1,9 @@
+package ExternallnterfaceTest;
+
 import Entities.Item;
 import ExternalInterface.ItemSearcher;
 import org.junit.jupiter.api.Assertions;
 import org.junit.jupiter.api.Test;
-
 import java.io.IOException;
 
 public class ItemSearchTest {
@@ -35,7 +36,7 @@ public class ItemSearchTest {
      * Testing search item by url feature and if it is scraping the correct value of price
      */
     @Test
-    public void searchItemTestKittenPlushPrice() throws IOException, InterruptedException {
+    public void searchItemTestKittenPlushPrice() throws IOException{
         ItemSearcher itemSearcher = new ItemSearcher();
 
         String kittenplush = "https://www.amazon.ca/Kitten-Plush-Stuffed-Animal-Pillow/dp/B088BWPFYZ/ref=sr_1_5?crid=16S7RSE5N25AA&keywords=uwu+cat&qid=1669082008&qu=eyJxc2MiOiIyLjI1IiwicXNhIjoiMC4wMCIsInFzcCI6IjAuMDAifQ%3D%3D&sprefix=uwu+cat%2Caps%2C69&sr=8-5";
@@ -50,7 +51,7 @@ public class ItemSearchTest {
      * Testing search item by url feature and if it is scraping the correct value of name
      */
     @Test
-    public void searchItemTestWorstPlayerToadName() throws IOException, InterruptedException {
+    public void searchItemTestWorstPlayerToadName() throws IOException {
         ItemSearcher itemSearcher = new ItemSearcher();
 
         String toadplush = "https://www.amazon.ca/uiuoutoy-Super-Mario-Bros-Mushroom/dp/B07CZGM7KV/ref=sr_1_1?crid=273IB3Y3TBWBZ&keywords=toad+mario+kart+plush&qid=1669082305&qu=eyJxc2MiOiIxLjQ1IiwicXNhIjoiMC4wMCIsInFzcCI6IjAuMDAifQ%3D%3D&s=books&sprefix=toad+mariokarplushie%2Cstripbooks%2C76&sr=1-1-catcorr";
@@ -64,7 +65,7 @@ public class ItemSearchTest {
      * Testing search item by url feature and if it is scraping the correct value of number of ratings for the product
      */
     @Test
-    public void searchItemCarnivalCountRatingTest() throws IOException, InterruptedException {
+    public void searchItemCarnivalCountRatingTest() throws IOException {
         ItemSearcher itemSearcher = new ItemSearcher();
 
         String carnivalplush = "https://www.amazon.ca/BB-Funhouse-Stuffed-Carnival-3-14inch/dp/B09NRGTT1R/ref=sr_1_7?crid=1QCCDI9NB8WIQ&keywords=whale+plush&qid=1669082056&qu=eyJxc2MiOiI2LjAyIiwicXNhIjoiNS4xMiIsInFzcCI6IjQuNTQifQ%3D%3D&sprefix=whale+plushie%2Caps%2C95&sr=8-7";
