+++ src/test/java/UseCasesTest/Notification/PriceDropNotificationTest.java
@@ -1,3 +1,5 @@
+package UseCasesTest.Notification;
+
 import UseCases.Notification.PriceDropNotification;
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.api.Assertions;
