+++ src/test/java/UseCasesTest/Notification/SaleNotificationTest.java
@@ -1,3 +1,5 @@
+package UseCasesTest.Notification;
+
 import UseCases.Notification.SaleNotification;
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.api.Assertions;
