+++ src/test/java/UseCasesTest/PriceHistory/PriceHistoryTest.java
@@ -1,13 +1,13 @@
+package UseCasesTest.PriceHistory;
+
 import UseCases.PriceHistory.GeneratePriceHistoryUseCase;
 import org.junit.jupiter.api.Assertions;
 import org.junit.jupiter.api.Test;
-
 import java.text.ParseException;
 import java.util.ArrayList;
 import java.util.Calendar;
 import java.util.Date;
 import Entities.*;
-
 import static java.util.Calendar.NOVEMBER;
